requirements.txt
altair==5.5.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.58.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
imbalanced-ensemble==0.2.3
imbalanced-learn==0.13.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
liac-arff==2.5.0
lightgbm==4.6.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
minio==7.2.15
narwhals==1.41.0
numba==0.61.2
numpy==2.2.6
openml==0.15.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pycryptodome==3.23.0
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
scikit-learn==1.6.0
scipy==1.15.3
seaborn==0.13.2
shap==0.47.2
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
smmap==5.0.2
streamlit==1.45.1
tdqm==0.0.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
xgboost==3.0.2
XlsxWriter==3.2.3
xmltodict==0.14.2
